requirements.txt
google-cloud-storage
requests
Flask